# swagger_server/swagger/swagger.yaml
---
swagger: "2.0"
info:
  description: "Sample version of DSF-SE API for communication with the professional\
    \ UI"
  version: "3"
  title: "DSF-SE API"
  contact:
    email: "sisay.adugna.chala@fit.fraunhofer.de"
host: "simplon.fit.fraunhofer.de"
basePath: "/se"
tags:
- name: "simulation"
  description: "Simulation"
- name: "commands"
  description: "Commands for the simulation engine"
schemes:
- "http"
paths:
  /simulation:
    post:
      tags:
      - "simulation"
      summary: "Send grid data to simulation engine in order to create a new simulation"
      operationId: "create_simulation"
      parameters:
      - in: "body"
        name: "body"
        description: "Grid to be simulated"
        required: true
        schema:
          $ref: "#/definitions/Grid"
      responses:
        200:
          description: "ID of the created simulation"
          schema:
            type: "string"
      x-swagger-router-controller: "swagger_server.controllers.simulation_controller"
  /simulation/{id}:
    get:
      tags:
      - "simulation"
      summary: "Get result of simulation"
      operationId: "get_simulation_result"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of the simulation"
        required: true
        type: "string"
      responses:
        200:
          description: "Results of successful simulation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/SimulationResult"
        400:
          description: "Invalid simulation configuration data supplied"
        404:
          description: "Simulation ID not found"
      x-swagger-router-controller: "swagger_server.controllers.simulation_controller"
    put:
      tags:
      - "simulation"
      summary: "Send new data to an existing simulation"
      operationId: "update_simulation"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of the simulation"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Updated grid data"
        required: true
        schema:
          $ref: "#/definitions/Grid"
      responses:
        default:
          description: "successful operation"
      x-swagger-router-controller: "swagger_server.controllers.simulation_controller"
    delete:
      tags:
      - "simulation"
      summary: "Delete a simulation and its data"
      operationId: "delete_simulation"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of the simulation"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
      x-swagger-router-controller: "swagger_server.controllers.simulation_controller"
  /commands/run/{id}:
    put:
      tags:
      - "commands"
      summary: "Runs a simulation"
      description: "Runs a simulation"
      operationId: "run_simulation"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of the simulation that should be started"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Configuration data for the simulation e.g. duration"
        required: false
        schema:
          $ref: "#/definitions/Simulation"
      responses:
        200:
          description: "Results of successful simulation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/SimulationResult"
        400:
          description: "Invalid simulation configuration data supplied"
        404:
          description: "Simulation ID not found"
      x-swagger-router-controller: "swagger_server.controllers.commands_controller"
  /commands/abort/{id}:
    put:
      tags:
      - "commands"
      summary: "Aborts a running simulation"
      description: "If the user of the professional GUI decides to abort a running\
        \ simulation this call will be triggered"
      operationId: "abort_simulation"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of the simulation that should be aborted"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
      x-swagger-router-controller: "swagger_server.controllers.commands_controller"
  /commands/status/{id}:
    get:
      tags:
      - "commands"
      summary: "Get the status of the simulation"
      operationId: "get_simulation_status"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of the simulation"
        required: true
        type: "string"
      responses:
        200:
          description: "Status of the simulation"
          schema:
            type: "number"
        400:
          description: "Invalid simulation configuration data supplied"
        404:
          description: "Simulation ID not found"
      x-swagger-router-controller: "swagger_server.controllers.commands_controller"
definitions:
  Grid:
    type: "object"
    properties:
      common:
        $ref: "#/definitions/Grid_Definition"
      radials:
        type: "array"
        items:
          $ref: "#/definitions/Radial"
  Grid_Definition:
    type: "object"
    required:
    - "MVAsc3"
    - "VoltageBases"
    - "angle"
    - "base_frequency"
    - "base_kV"
    - "bus1"
    - "id"
    - "phases"
    properties:
      id:
        type: "string"
      base_kV:
        type: "number"
      per_unit:
        type: "number"
      phases:
        type: "integer"
      bus1:
        type: "string"
      angle:
        type: "integer"
      MVAsc3:
        type: "number"
      MVAsc1:
        type: "number"
      base_frequency:
        type: "integer"
        format: "int32"
      VoltageBases:
        type: "array"
        items:
          type: "number"
      url_storage_controller:
        type: "string"
        default: "http://localhost:8080"
      city:
        type: "string"
      country:
        type: "string"
  Radial:
    type: "object"
    properties:
      radialId:
        type: "string"
      transformer:
        type: "array"
        items:
          $ref: "#/definitions/Transformer"
      loads:
        type: "array"
        items:
          $ref: "#/definitions/Load"
      powerProfiles:
        type: "array"
        items:
          $ref: "#/definitions/PowerProfile"
      powerLines:
        type: "array"
        items:
          $ref: "#/definitions/Powerline"
      photovoltaics:
        type: "array"
        items:
          $ref: "#/definitions/Photovoltaic"
      storageUnits:
        type: "array"
        items:
          $ref: "#/definitions/Storage"
      chargingPoints:
        type: "array"
        items:
          $ref: "#/definitions/ChargingPoint"
      linecode:
        type: "array"
        items:
          $ref: "#/definitions/Linecode"
      capacitor:
        type: "array"
        items:
          $ref: "#/definitions/Capacitor"
      regcontrol:
        type: "array"
        items:
          $ref: "#/definitions/RegControl"
      xycurves:
        type: "array"
        items:
          $ref: "#/definitions/XYCurve"
      loadshapes:
        type: "array"
        items:
          $ref: "#/definitions/Loadshape"
      tshapes:
        type: "array"
        items:
          $ref: "#/definitions/Tshape"
  Node:
    type: "object"
    required:
    - "id"
    properties:
      id:
        type: "string"
    example:
      id: "id"
  Phases:
    type: "object"
    required:
    - "phase_R"
    - "phase_S"
    - "phase_T"
    properties:
      phase_R:
        type: "boolean"
      phase_S:
        type: "boolean"
      phase_T:
        type: "boolean"
  RegControl:
    type: "object"
    required:
    - "id"
    properties:
      id:
        type: "string"
      transformer:
        type: "string"
        description: "ID of the transformer to be connected to"
      winding:
        type: "integer"
      vreg:
        type: "number"
      band:
        type: "number"
      ptration:
        type: "number"
      ctprim:
        type: "number"
      r:
        type: "number"
      x:
        type: "number"
  Capacitor:
    type: "object"
    required:
    - "bus"
    - "id"
    - "kV"
    - "kVAR"
    - "phases"
    properties:
      id:
        type: "string"
      bus:
        type: "string"
        description: "ID for the connected node"
      phases:
        type: "integer"
      kVAR:
        type: "number"
      kV:
        type: "number"
  Load:
    type: "object"
    required:
    - "bus"
    - "id"
    - "kV"
    - "kW"
    - "phases"
    properties:
      id:
        type: "string"
      bus:
        type: "string"
        description: "ID for the connected node"
      phases:
        type: "integer"
      connection_type:
        type: "string"
      model:
        type: "integer"
        description: "Defining load models for openDSS"
        minimum: 1
        maximum: 8
      kV:
        type: "number"
      kW:
        type: "number"
      kVar:
        type: "number"
  PowerProfile:
    type: "object"
    properties:
      id:
        type: "string"
      items:
        type: "number"
  Powerline:
    type: "object"
    required:
    - "bus1"
    - "bus2"
    - "id"
    - "length"
    - "phases"
    properties:
      id:
        type: "string"
      bus1:
        type: "string"
        description: "ID for the connected node"
      phases:
        type: "integer"
      bus2:
        type: "string"
        description: "ID for the connected node"
      length:
        type: "number"
      unitlength:
        type: "string"
      linecode:
        type: "string"
      r1:
        type: "number"
      r0:
        type: "number"
      x1:
        type: "number"
      x0:
        type: "number"
      c1:
        type: "number"
      c0:
        type: "number"
      switch:
        type: "boolean"
        default: false
  Linecode:
    type: "object"
    required:
    - "id"
    - "nphases"
    - "rmatrix"
    - "units"
    - "xmatrix"
    properties:
      id:
        type: "string"
      nphases:
        type: "number"
      rmatrix:
        type: "array"
        items:
          type: "array"
          items:
            type: "number"
      xmatrix:
        type: "array"
        items:
          type: "array"
          items:
            type: "number"
      cmatrix:
        type: "array"
        items:
          type: "array"
          items:
            type: "number"
      units:
        type: "string"
  XYCurve:
    type: "object"
    required:
    - "id"
    - "npts"
    - "xarray"
    - "yarray"
    properties:
      id:
        type: "string"
      npts:
        type: "number"
      xarray:
        type: "array"
        items:
          type: "number"
      yarray:
        type: "array"
        items:
          type: "number"
  Loadshape:
    type: "object"
    required:
    - "id"
    - "interval"
    - "mult"
    - "npts"
    properties:
      id:
        type: "string"
      npts:
        type: "number"
      interval:
        type: "number"
      mult:
        type: "array"
        items:
          type: "number"
  Tshape:
    type: "object"
    required:
    - "id"
    - "interval"
    - "npts"
    - "temp"
    properties:
      id:
        type: "string"
      npts:
        type: "number"
      interval:
        type: "number"
      temp:
        type: "array"
        items:
          type: "number"
  Storage:
    type: "object"
    required:
    - "bus1"
    - "id"
    - "kv"
    - "optimization_model"
    - "storage_capacity"
    properties:
      id:
        type: "string"
      bus1:
        type: "string"
        description: "ID for the connected bus"
      phases:
        type: "integer"
      connection:
        type: "string"
        description: "={wye|LN|delta|LL}. Default is wye. (openDSS Name: conn)"
      soc:
        type: "number"
        description: "State of Charge in % of rated kWh (openDSS Name: %stored)"
      kv:
        type: "number"
        description: "Nominal rated (1.0 per unit) voltage, kV, for Storage element."
      kw_rated:
        type: "number"
        description: "kW rating of power output. Side effect: Set KVA property."
      storage_capacity:
        type: "number"
        description: "Rated storage capacity in kWh. Default is 50."
      kwh_stored:
        type: "number"
        description: "Present amount of energy stored, kWh. Default is same as kWh\
          \ rated."
      max_charging_power:
        type: "number"
        description: "Maximal power to charge the storage"
      max_discharging_power:
        type: "number"
        description: "Maximal power to discharge the storage"
      charge_efficiency:
        type: "number"
        description: "Efficiency for CHARGING the storage element. Default=90%. (%EffCharge)"
      discharge_efficiency:
        type: "number"
        description: "Efficiency for DISCHARGING the storage element. Default=90%.\
          \ Idling losses are handled by %IdlingkW. (%EffCharge)"
      powerfactor:
        type: "number"
        description: "(openDSS Name: pf)"
      optimization_model:
        type: "string"
        description: "(A value required by PROFFES)"
      global_control:
        type: "boolean"
  Photovoltaic:
    type: "object"
    required:
    - "bus1"
    - "id"
    - "kV"
    - "max_power_kW"
    - "phases"
    properties:
      id:
        type: "string"
      inverterEfficiency:
        type: "number"
      phases:
        type: "integer"
      power:
        type: "number"
      max_power_kW:
        type: "number"
      powerunit:
        type: "string"
      powerfactor:
        type: "number"
      bus1:
        type: "string"
      kV:
        type: "number"
      effcurve:
        type: "string"
      ptcurve:
        type: "string"
      daily:
        type: "string"
      tdaily:
        type: "string"
      temperature:
        type: "number"
      irrad:
        type: "number"
      voltageunit:
        type: "string"
  ChargingPoint:
    type: "object"
    required:
    - "id"
    - "node"
    properties:
      id:
        type: "string"
      node:
        type: "string"
        description: "ID for the connected node"
    example:
      node: "node"
      id: "id"
  Winding:
    type: "object"
    properties:
      wdg:
        type: "integer"
        description: "Integer representing the winding witch will become the active\
          \ winding for subsequent data"
      bus:
        type: "string"
        description: "Definition for the connector of this winding (each winding is\
          \ connected to one terminal of the transformator and, hence, to one bus"
      connection_type:
        type: "string"
      kV:
        type: "number"
        description: "Rated voltage of this winding, kV"
      kVA:
        type: "number"
        description: "base kVA rating for windings"
      percent_resistance:
        type: "number"
      xlt:
        type: "integer"
      xht:
        type: "integer"
    example:
      wdg: 1
      bus: "SourceBus"
      connection_type: "wye"
      kV: 4.160000000000000142108547152020037174224853515625
      kVA: 5000
      percent_resistance: 0.5
  Transformer:
    type: "object"
    required:
    - "buses"
    - "id"
    - "kvas"
    - "kvs"
    - "phases"
    - "windings"
    properties:
      id:
        type: "string"
      phases:
        type: "integer"
        description: "Number of Phases"
      windings:
        type: "integer"
        description: "number of windings"
      buses:
        type: "array"
        description: "Array of bus definitions for windings [1, 2. …]."
        items:
          type: "string"
      kvas:
        type: "array"
        description: "Array of base kVA ratings for windings [1,2,…]."
        items:
          type: "number"
      kvs:
        type: "array"
        description: "Array of kV ratings for the kV field for windings [1,2,…]."
        items:
          type: "number"
      conns:
        type: "array"
        description: "Array of winding connections for windings [1, 2. …] <wye|delta>."
        items:
          type: "string"
      xsc_array:
        type: "array"
        description: "Array of n*(n-1) /2 short circuit reactances in percent on the\
          \ first winding’s kVA base. “n” is the number of windings. Order (12, 13,\
          \ 14, …1n, 23, 24, … 34, …)"
        items:
          type: "number"
      percent_rs:
        type: "array"
        description: "Array of percent resistances for windings [1, 2. …]"
        items:
          type: "number"
      percent_load_loss:
        type: "number"
        description: "Percent Losses at rated load.. Causes the %r values to be set\
          \ for windings 1 and 2."
      bank:
        type: "string"
        description: "Name of the bank this transformer is part of"
      taps:
        type: "array"
        description: "Array of per unit taps for windings [1,2,…]."
        items:
          type: "number"
          minimum: 0
          maximum: 1
  Duration:
    type: "object"
    properties:
      day:
        type: "integer"
        format: "int32"
      month:
        type: "integer"
        format: "int32"
      year:
        type: "integer"
        format: "int32"
    example:
      month: 6
      year: 1
      day: 0
  Simulation:
    type: "object"
    required:
    - "sim_duration_in_days"
    properties:
      sim_duration_in_days:
        type: "integer"
        format: "int32"
      thresholdLow:
        type: "number"
        format: "float"
        description: "Lower threshold in % for voltage error"
        minimum: 0
        maximum: 100
      thresholdMedium:
        type: "number"
        format: "float"
        description: "Medium threshold in % for voltage error"
        minimum: 0
        maximum: 100
      thresholdHigh:
        type: "number"
        format: "float"
        description: "Maximum threshold in % for voltage error"
        minimum: 0
        maximum: 100
    example:
      sim_duration_in_days: 365
      thresholdMedium: 5
      thresholdHigh: 10
      thresholdLow: 2.5
  SimulationResult:
    type: "object"
    properties:
      nodeId:
        type: "string"
      voltage:
        $ref: "#/definitions/Voltage"
      error:
        $ref: "#/definitions/Error"
    example:
      error:
        voltageT: true
        voltageR: true
        voltageS: true
      nodeId: "nodeId"
      voltage:
        voltageT: 1.46581298050294517310021547018550336360931396484375
        voltageR: 0.80082819046101150206595775671303272247314453125
        voltageS: 6.02745618307040320615897144307382404804229736328125
  Voltage:
    type: "object"
    properties:
      voltageR:
        type: "number"
      voltageS:
        type: "number"
      voltageT:
        type: "number"
    example:
      voltageT: 1.46581298050294517310021547018550336360931396484375
      voltageR: 0.80082819046101150206595775671303272247314453125
      voltageS: 6.02745618307040320615897144307382404804229736328125
  Error:
    type: "object"
    properties:
      voltageR:
        type: "boolean"
      voltageS:
        type: "boolean"
      voltageT:
        type: "boolean"
    example:
      voltageT: true
      voltageR: true
      voltageS: true
  Transformer_wdgs:
    properties:
      struct:
        type: "array"
        items:
          $ref: "#/definitions/Winding"
  Transformer_regcontrol:
    properties:
      struct:
        type: "array"
        items:
          $ref: "#/definitions/RegControl"
externalDocs:
  description: "Find out more about Storage4Grid"
  url: "http://www.storage4grid.eu/pages/index.html"